move textspan to different namespace
