+++ modules/Code/Syntax/SyntaxNode.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.IO;
+using swifty.Code.Text;
 
 namespace swifty.Code.Syntaxt {
     public abstract class SyntaxNode {
