+++ modules/Code/Syntax/SyntaxToken.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using swifty.Code.Text;
 using System.Linq;
 
 namespace swifty.Code.Syntaxt {
